Fallback to UTC if no managed customers are available
Motivation
----------

https://stitchdata.atlassian.net/browse/SUP-292

For context, we [previously][4] fixed parsing of `startDate` and `endDate`
fields on `campaign`s to parse through the parent account's timezone into
UTC as `YYYYMMDD` date strings. This remains in accordance with our best
understanding of the nature of those two fields based on the [Campaign API
Docs][5].

[4]: https://github.com/timvisher/tap-adwords/commit/20a02fa04a5ad0a39fcd85b7a0d2b9e728902a54
[5]: https://developers.google.com/adwords/api/docs/reference/v201809/CampaignService.Campaign#startdate

This worked for most customers but for some it resulted in an error
because the request for their account resulted in an empty resultset.

Based on my understanding of the [API Docs][1] there should be no reason
why a [ManagedCustomer][2] shouldn't come back from the
[ManagedCustomerService][3] for a customer id that the client has access
to. Nevertheless we've seen multiple occasions where none come back.
Falling back to UTC seems like a reasonable default although it's almost
certainly wrong.

[1]: https://developers.google.com/adwords/api/docs/reference/v201809/ManagedCustomerService
[2]: https://developers.google.com/adwords/api/docs/reference/v201809/ManagedCustomerService.ManagedCustomer
[3]: https://developers.google.com/adwords/api/docs/reference/v201809/ManagedCustomerService#get